+++ src/terminal_ghosttyP.h
@@ -44,6 +44,9 @@ struct XtpTerminal
         bool reverse_colors;
         bool bold_colors;
         bool allow_color_ops;
+        bool allow_mouse_ops;
+        bool allow_tcap_ops;
+        XtpTcapOps tcap_ops;
         XtpColorOps color_ops;
         bool color_list_active;
         bool color_list_skipping;